lambda_functions/v1/requirements/dev-requirements.txt
pytest
flake8
flake8-quotes
Flask
connexion
pytest-env
boto3
pytest-cov
requests-aws4auth
pytest-html
moto
pyjwt
hypothesis
validators
yarl
localstack-client
tenacity
redis
fakeredis
